Ensure linked suproject does not delete target

+++ trans/models/__init__.py
@@ -40,5 +40,11 @@ def delete_object_dir(sender, instance, **kwargs):
     Handler to delete (sub)project directory on project deletion.
     '''
     project_path = instance.get_path()
+
+    # Do not delete liked subprojects
+    if hasattr(instance, 'is_repo_link') and instance.is_repo_link():
+        return
+
+    # Remove path if it exists
     if os.path.exists(project_path):
         shutil.rmtree(project_path)
